requirements.txt
Python==3.9.6
Flask==2.0.1
Flask-SocketIO==5.1.1
Flask-JWT-Extended==4.3.1
mysql-connector-python==8.0.27
